loaders(quarkus): consistent test class naming
