+++ camel-k-quarkus/camel-k-quarkus-itests/camel-k-quarkus-itests-loader-groovy/src/test/java/org/apache/camel/k/loader/yaml/quarkus/GroovyLoaderTest.java
@@ -14,13 +14,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.k.loader.js.quarkus;
+package org.apache.camel.k.loader.yaml.quarkus;
 
 import java.io.IOException;
 import java.io.InputStream;
 
 import javax.ws.rs.core.MediaType;
 
+import io.quarkus.test.junit.DisabledOnNativeImage;
 import io.quarkus.test.junit.QuarkusTest;
 import io.restassured.RestAssured;
 import io.restassured.path.json.JsonPath;
@@ -29,13 +30,14 @@ import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+@DisabledOnNativeImage
 @QuarkusTest
-public class ExtensionTest {
+public class GroovyLoaderTest {
     @Test
     public void testLoadRoutes() throws IOException {
         String code;
 
-        try (InputStream is = ExtensionTest.class.getResourceAsStream("/routes.js")) {
+        try (InputStream is = GroovyLoaderTest.class.getResourceAsStream("/routes.groovy")) {
             code = IOHelper.loadText(is);
         }
 
@@ -51,7 +53,7 @@ public class ExtensionTest {
                 .jsonPath();
 
         assertThat(p.getList("components", String.class)).contains("direct", "log");
-        assertThat(p.getList("routes", String.class)).contains("js");
-        assertThat(p.getList("endpoints", String.class)).contains("direct://js", "log://js");
+        assertThat(p.getList("routes", String.class)).contains("groovy");
+        assertThat(p.getList("endpoints", String.class)).contains("direct://groovy", "log://groovy");
     }
 }
